Simplify tests.
